+++ tests/Methods/EncodingTest.php
@@ -1,6 +1,4 @@
 <?php
-
-/** @noinspection UnknownInspectionInspection */
 declare(strict_types=1);
 
 namespace Tests\Methods;